comment, move proto utils to main lib
