--- lib/src/main/java/grpcbridge/route/ManifestGenerator.java
@@ -1,7 +1,0 @@
-package grpcbridge.route;
-
-import java.util.List;
-
-public interface ManifestGenerator {
-    String generate(List<Route> routes);
-}
