+++ swagger/src/main/java/grpcbridge/swagger/ModelBuilder.java
@@ -6,8 +6,8 @@ import com.google.protobuf.Descriptors.FieldDescriptor;
 import grpcbridge.swagger.model.Property;
 import grpcbridge.swagger.model.EnumSwaggerModel;
 import grpcbridge.swagger.model.SwaggerModel;
-import grpcbridge.swagger.proto.ProtoDescriptorTraverser;
-import grpcbridge.swagger.proto.ProtoVisitor;
+import grpcbridge.util.ProtoDescriptorTraverser;
+import grpcbridge.util.ProtoVisitor;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Stack;
