+++ swagger/src/main/java/grpcbridge/swagger/ParametersBuilder.java
@@ -12,8 +12,8 @@ import grpcbridge.swagger.model.Parameter;
 import grpcbridge.swagger.model.Parameter.Location;
 import grpcbridge.swagger.model.Property;
 import grpcbridge.swagger.model.SwaggerModel;
-import grpcbridge.swagger.proto.ProtoDescriptorTraverser;
-import grpcbridge.swagger.proto.ProtoVisitor;
+import grpcbridge.util.ProtoDescriptorTraverser;
+import grpcbridge.util.ProtoVisitor;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
